requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.3.0
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-crontab==0.7.1
django-oauth-toolkit==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.0.0
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2==2.9.6
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
stripe==5.4.0
typing_extensions==4.5.0
urllib3==1.26.12
wrapt==1.15.0
